added audio, update main menu
menu works, audio is added to menu, need to link up the settings to actually change the audio / sfx volume. credits display over the controls. game closes when exit is pressed. need to add a way to close the game from a build / pause menu with a close game option